Api-Bancaria/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
